split dwi preproc tests
